Moving admin methods to its own controllers.
