+++ src/Jobtech.OpenPlatforms.GigDataApi.Api/Controllers/PlatformController.cs
@@ -56,189 +56,6 @@ namespace Jobtech.OpenPlatforms.GigDataApi.Api.Controllers
         }
 
         /// <summary>
-        /// Create a new platform
-        /// </summary>
-        /// <remarks>
-        /// Admin method
-        /// </remarks>
-        /// <param name="adminKey">The admin key</param>
-        /// <param name="model">The data needed for creating a new platform</param>
-        /// <param name="cancellationToken"></param>
-        /// <returns></returns>
-        [AllowAnonymous]
-        //[ApiExplorerSettings(IgnoreApi = false)]
-        [HttpPost("admin/create")]
-        public async Task<ActionResult<PlatformInfoViewModel>> CreatePlatform(
-            [FromHeader(Name = "admin-key")] Guid adminKey, [FromBody] CreatePlatformModel model,
-            CancellationToken cancellationToken)
-        {
-            if (!_options.AdminKeys.Contains(adminKey))
-            {
-                return Unauthorized();
-            }
-
-            using var session = _documentStore.OpenAsyncSession();
-            var createdPlatform = await _platformManager.CreatePlatform(model.Name, model.AuthMechanism,
-                PlatformIntegrationType.GigDataPlatformIntegration,
-                new RatingInfo(model.MinRating, model.MaxRating, model.RatingSuccessLimit),
-                3600, Guid.NewGuid(), model.Description, model.LogoUrl, session, true, cancellationToken);
-
-            await session.SaveChangesAsync(cancellationToken);
-
-            return new PlatformInfoViewModel(createdPlatform.ExternalId, createdPlatform.Name,
-                createdPlatform.Description, createdPlatform.LogoUrl, createdPlatform.IsInactive, createdPlatform.AuthenticationMechanism);
-        }
-
-        /// <summary>
-        /// Get info about platform with the given platform id.
-        /// </summary>
-        /// <remarks>
-        /// Admin method
-        /// </remarks>
-        /// <param name="adminKey">The admin key</param>
-        /// <param name="platformId">The platform id</param>
-        /// <param name="cancellationToken"></param>
-        /// <returns></returns>
-        [AllowAnonymous]
-        //[ApiExplorerSettings(IgnoreApi = false)]
-        [HttpGet("{platformId}/admin/info")]
-        public async Task<ActionResult<PlatformInfoViewModel>> GetPlatformInfo(
-            [FromHeader(Name = "admin-key")] Guid adminKey, Guid platformId, CancellationToken cancellationToken)
-        {
-            if (!_options.AdminKeys.Contains(adminKey))
-            {
-                return Unauthorized();
-            }
-
-            using var session = _documentStore.OpenAsyncSession();
-            var platform = await _platformManager.GetPlatformByExternalId(platformId, session, cancellationToken);
-            return new PlatformInfoViewModel(platform.ExternalId, platform.Name, platform.Description, platform.LogoUrl,
-                platform.IsInactive, platform.AuthenticationMechanism);
-        }
-
-        /// <summary>
-        /// Set platform to active.
-        /// </summary>
-        /// <remarks>
-        /// Admin method
-        /// </remarks>
-        /// <param name="adminKey">The admin key</param>
-        /// <param name="platformId">The platform id</param>
-        /// <param name="cancellationToken"></param>
-        /// <returns></returns>
-        [AllowAnonymous]
-        //[ApiExplorerSettings(IgnoreApi = false)]
-        [HttpPatch("{platformId}/admin/activate")]
-        public async Task<ActionResult> ActivatePlatform([FromHeader(Name = "admin-key")] Guid adminKey,
-            Guid platformId, CancellationToken cancellationToken)
-        {
-            if (!_options.AdminKeys.Contains(adminKey))
-            {
-                return Unauthorized();
-            }
-
-            using var session = _documentStore.OpenAsyncSession();
-            var platform = await _platformManager.GetPlatformByExternalId(platformId, session, cancellationToken);
-            platform.IsInactive = false;
-
-            await session.SaveChangesAsync(cancellationToken);
-
-            return Ok("Platform set to active");
-        }
-
-        /// <summary>
-        /// Set platform to inactive
-        /// </summary>
-        /// <remarks>
-        /// Admin method
-        /// </remarks>
-        /// <param name="adminKey">The admin key</param>
-        /// <param name="platformId">The platform id</param>
-        /// <param name="cancellationToken"></param>
-        /// <returns></returns>
-        [AllowAnonymous]
-        //[ApiExplorerSettings(IgnoreApi = false)]
-        [HttpPatch("{platformId}/admin/deactivate")]
-        public async Task<ActionResult> DeactivatePlatform([FromHeader(Name = "admin-key")] Guid adminKey,
-            Guid platformId, CancellationToken cancellationToken)
-        {
-            if (!_options.AdminKeys.Contains(adminKey))
-            {
-                return Unauthorized();
-            }
-
-            using var session = _documentStore.OpenAsyncSession();
-            var platform = await _platformManager.GetPlatformByExternalId(platformId, session, cancellationToken);
-            platform.IsInactive = true;
-
-            await session.SaveChangesAsync(cancellationToken);
-
-            return Ok("Platform set to inactive");
-        }
-
-        /// <summary>
-        /// Set the logo url for a platform
-        /// </summary>
-        /// <remarks>
-        /// Admin method
-        /// </remarks>
-        /// <param name="adminKey">The admin key</param>
-        /// <param name="platformId">The platform id</param>
-        /// <param name="model">Logo url data</param>
-        /// <param name="cancellationToken"></param>
-        /// <returns></returns>
-        [AllowAnonymous]
-        //[ApiExplorerSettings(IgnoreApi = false)]
-        [HttpPatch("{platformId}/admin/set-logourl")]
-        public async Task<ActionResult> SetLogoUrl([FromHeader(Name = "admin-key")] Guid adminKey, Guid platformId,
-            [FromBody] SetLogoUrlModel model, CancellationToken cancellationToken)
-        {
-            if (!_options.AdminKeys.Contains(adminKey))
-            {
-                return Unauthorized();
-            }
-
-            using var session = _documentStore.OpenAsyncSession();
-            var platform = await _platformManager.GetPlatformByExternalId(platformId, session, cancellationToken);
-            platform.LogoUrl = model.LogoUrl;
-
-            await session.SaveChangesAsync(cancellationToken);
-
-            return Ok("Platform logo url updated");
-        }
-
-        /// <summary>
-        /// Set the description for a platform
-        /// </summary>
-        /// <remarks>
-        /// Admin method
-        /// </remarks>
-        /// <param name="adminKey">The admin key</param>
-        /// <param name="platformId">The platform id</param>
-        /// <param name="model">The description data</param>
-        /// <param name="cancellationToken"></param>
-        /// <returns></returns>
-        [AllowAnonymous]
-        //[ApiExplorerSettings(IgnoreApi = false)]
-        [HttpPatch("{platformId}/admin/set-description")]
-        public async Task<ActionResult> SetDescription([FromHeader(Name = "admin-key")] Guid adminKey, Guid platformId,
-            [FromBody] SetDescriptionModel model, CancellationToken cancellationToken)
-        {
-            if (!_options.AdminKeys.Contains(adminKey))
-            {
-                return Unauthorized();
-            }
-
-            using var session = _documentStore.OpenAsyncSession();
-            var platform = await _platformManager.GetPlatformByExternalId(platformId, session, cancellationToken);
-            platform.Description = model.Description;
-
-            await session.SaveChangesAsync(cancellationToken);
-
-            return Ok("Platform logo url updated");
-        }
-
-        /// <summary>
         /// Get the connection status for a given platform.
         /// </summary>
         /// <remarks>
@@ -621,18 +438,6 @@ namespace Jobtech.OpenPlatforms.GigDataApi.Api.Controllers
         }
 
         public bool IsConnected { get; set; }
-    }
-
-    public class PlatformInfoViewModel : PlatformViewModel
-    {
-        public PlatformInfoViewModel(Guid externalPlatformId, string name, string description, string logoUrl,
-            bool isInactive, PlatformAuthenticationMechanism authMechanism) : base(externalPlatformId, name,
-            description, logoUrl, authMechanism)
-        {
-            IsInactive = isInactive;
-        }
-
-        public bool IsInactive { get; set; }
     }
 
     public class AppUserConnectionViewModel
